+++ src/pages/About/About.tsx
@@ -7,9 +7,18 @@ import { JsonSection } from '../../components/layout/JsonSection/JsonSection';
 import { PageTile } from '../../components/layout/ResponsiveTile/PageTile';
 import { colors } from '../../constants/colors';
 import {
+	gapYah2Link,
+	gapYahLink,
+	kotlinLink,
+	linkedinProfileLink,
+	litterAppLink,
+	recylotronLink,
+	sdlcLink,
+	vb6Link,
+} from '../../constants/link-constants';
+import {
 	litterImages,
 	litterTileImage,
-	panelData,
 	recyclotronImages,
 	recyclotronTileImage,
 	uniBackground,
@@ -19,22 +28,43 @@ import {
 } from './about-constants';
 import './About.css';
 
+const vb6 = `<@${vb6Link.text}@${vb6Link.link}@${vb6Link.img}@>`;
+const sdlc = `<@${sdlcLink.text}@${sdlcLink.link}@>`;
+const gapYah = `<@${gapYahLink.text}@${gapYahLink.link}@>`;
+const gapYah2 = `<@${gapYah2Link.text}@${gapYah2Link.link}@>`;
+const recylotron = `<@${recylotronLink.text}@${recylotronLink.link}@${recylotronLink.img}@>`;
+const litterApp = `<@${litterAppLink.text}@${litterAppLink.link}@${litterAppLink.img}@>`;
+const kotlin = `<@${kotlinLink.text}@${kotlinLink.link}@${kotlinLink.img}@>`;
+const linkedin = `<@${linkedinProfileLink.text}@${linkedinProfileLink.link}@@>`;
+
 const blurbPanel: JsonSectionPanel = {
 	header: {
-		titleText: panelData.school.title,
-		subtitleText: panelData.school.subtitle,
+		titleText: 'about',
+		subtitleText: '(how I became a Software Engineer)',
 	},
 	imageSlot: {
 		image: vb6IdeImage,
 	},
 	imageMaxWidth: '40%',
-	content: panelData.school.content,
+	content: [
+		`My first real taste of the ${sdlc} was in school.`,
+		`Using ${vb6} I built a bookkeeping app for a local book shop 📚`,
+		'',
+		'I recall spending days to add currency functionality, writing dozens of if-statements...',
+		'',
+		'Thankfully, that code has been lost to time.',
+		'',
+		'But it inspired me to switch my university application from Chemistry to Computer Science.',
+		"This meant next year was getting my Maths A-level (which I'd skipped) to meet course requirements...",
+		'',
+		`...so my ${gapYah} ${gapYah2} was spent learning 📐 trigonometry...!`,
+	],
 	contentBackground: colors.rust,
 };
 
 const uniPanel: JsonSectionPanel = {
 	header: {
-		titleText: panelData.uni.title,
+		titleText: 'University',
 	},
 	imageSlot: {
 		images: uniImages,
@@ -45,7 +75,20 @@ const uniPanel: JsonSectionPanel = {
 	imageMaxWidth: '35%',
 	textMinWidth: '50%',
 	minHeight: '60vh',
-	content: panelData.uni.content,
+	content: [
+		'',
+		"It wasn't until third year, I started enjoying uni.",
+		'I had fumbled my first two years, having to resit multiple exams.',
+		'I knuckled down, keen for the classes I had chosen.',
+		'',
+		'Then, come March 2020 - everything stopped. I flew back home, unsure of the future.',
+		'When obvious this was no Spring thing, I cancelled the flat lease.',
+		"I wouldn't return to Edinburgh until 2 years later.",
+		'',
+		'It was a good time to finish my degree - I had swapped socialising for studying anyway.',
+		'Besides, the Informatics department was (mostly) well equipped to virtually assess.',
+		'',
+	],
 	contentBackground: colors.rust,
 };
 
@@ -54,8 +97,20 @@ const recyclotronPanel: JsonSectionPanel = {
 		images: recyclotronImages,
 		cyclerInterval: 10000,
 	},
+	mobileOrder: 'text-first',
 	imageMaxWidth: '35%',
-	content: panelData.recyclotron.content,
+	content: [
+		'',
+		"There was also a 'build-us-a-robot' course.",
+		'Our group decided to make an auto-sorting bin.',
+		'',
+		'We trained a neural-network AI on images of rubbish to classify waste into categories.',
+		'The rubbish would enter a chamber which would analyse the properties of the material.',
+		'It would then be sorted into a different bin to recycle.',
+		'',
+		`You can learn more on ${recylotron} project page.`,
+		'',
+	],
 	contentBackground: colors.rust,
 	contentChildren: (
 		<div
@@ -78,7 +133,16 @@ const litterPanel: JsonSectionPanel = {
 	imageMaxWidth: '30%',
 	minHeight: '80vh',
 	reverseColumns: true,
-	content: panelData.litter.content,
+	content: [
+		'',
+		'My final year project was a software engineering project.',
+		`I was to create a Location-based Social Media app, for Android, using ${kotlin}.`,
+		'',
+		'It felt exactly like the school project that inspired the degree 5 years earlier...',
+		'',
+		`You can learn more on the ${litterApp} project page.`,
+		'',
+	],
 	contentBackground: colors.rust,
 	contentChildren: (
 		<div
@@ -94,7 +158,13 @@ const litterPanel: JsonSectionPanel = {
 };
 
 const jobPanel: JsonTextPanelData = {
-	content: panelData.job.content,
+	content: [
+		'',
+		'I still try to scratch my creative itch outside of my day job...',
+		'',
+		`You can visit my ${linkedin} page for info on my career projects.`,
+		'',
+	],
 	contentBackground: colors.teal,
 	maxWidth: '50%',
 };
@@ -119,7 +189,13 @@ function About() {
 					jobPanel,
 				]}
 			/>
-			{/* 
+		</>
+	);
+}
+
+//TODO - where should these recommendations live?
+{
+	/* 
 			<div className='about-image-text-section'>
 				<TileList
 					items={recommendations.books}
@@ -173,10 +249,10 @@ function About() {
 					<p>{'hey'}</p>
 					<p>{'I make things, like this website'}</p>
 					<p>{'ye'}</p>
-				</ImageTextLayout> */}
-			{/* </div> */}
-		</>
-	);
+				</ImageTextLayout> */
+}
+{
+	/* </div> */
 }
 
 export { About };
