--- src/pages/Projects/LitterApp/Litter.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import { PageInConstruction } from '../../../components/display/PageInConstruction/PageInConstruction';
-
-function LitterApp() {
-	return (
-		<div className='App'>
-			{' '}
-			<PageInConstruction></PageInConstruction>
-		</div>
-	);
-}
-
-export { LitterApp };
