+++ src/pages/Projects/LitterApp/litterapp-constants.ts
@@ -1,0 +1,36 @@
+
+const blurb = 'LitterApp, Location-based Social Media';
+
+const panelData = {
+	precursor: {
+		title: 'Informatics Large Practical',
+		content: [
+			'',
+		],
+	},
+	software: {
+		title: 'Software',
+		content: [
+			'',
+		],
+	},
+	ai: {
+		title: 'AI',
+		content: [
+			'',
+		],
+	},
+
+	hardware: {
+		title: 'Hardware',
+		subtitle: blurb,
+		content: [
+			'',
+		],
+	},
+};
+
+export {
+	blurb,
+	panelData,
+};
