+++ src/pages/Projects/Recylotron/recylotron-constants.ts
@@ -1,0 +1,36 @@
+
+const blurb = 'Recylotron, Smart-Bin';
+
+const panelData = {
+	intro: {
+		title: 'What?',
+		content: [
+			'',
+		],
+	},
+	software: {
+		title: 'Software',
+		content: [
+			'',
+		],
+	},
+	ai: {
+		title: 'AI',
+		content: [
+			'',
+		],
+	},
+
+	hardware: {
+		title: 'Hardware',
+		subtitle: blurb,
+		content: [
+			'',
+		],
+	},
+};
+
+export {
+	blurb,
+	panelData,
+};
